+++ ogs5py/fileclasses/bc/__init__.py
@@ -12,8 +12,6 @@ File Class
 
 ----
 """
-from __future__ import absolute_import
-
 from ogs5py.fileclasses.bc.core import BC
 
 __all__ = ["BC"]