+++ ogs5py/fileclasses/fct/__init__.py
@@ -12,8 +12,6 @@ File Class
 
 ----
 """
-from __future__ import absolute_import
-
 from ogs5py.fileclasses.fct.core import FCT
 
 __all__ = ["FCT"]