+++ ogs5py/fileclasses/ic/__init__.py
@@ -13,8 +13,6 @@ File Classes
 
 ----
 """
-from __future__ import absolute_import
-
 from ogs5py.fileclasses.ic.core import IC, RFR
 
 __all__ = ["IC", "RFR"]