+++ ogs5py/fileclasses/mmp/__init__.py
@@ -12,8 +12,6 @@ File Class
 
 ----
 """
-from __future__ import absolute_import
-
 from ogs5py.fileclasses.mmp.core import MMP
 
 __all__ = ["MMP"]