+++ ogs5py/fileclasses/msh/__init__.py
@@ -19,8 +19,6 @@ File Class
 
 ----
 """
-from __future__ import absolute_import
-
 from ogs5py.fileclasses.msh.core import MSH
 
 __all__ = ["MSH"]