+++ ogs5py/fileclasses/msh/checker.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 """Checker for the ogs5py mesh file."""
-from __future__ import division, print_function, absolute_import
 import numpy as np
 from ogs5py.fileclasses.msh.tools import no_of_elements
 from ogs5py.tools.types import (