+++ ogs5py/fileclasses/out/__init__.py
@@ -12,8 +12,6 @@ File Class
 
 ----
 """
-from __future__ import absolute_import
-
 from ogs5py.fileclasses.out.core import OUT
 
 __all__ = ["OUT"]