+++ ogs5py/fileclasses/pcs/__init__.py
@@ -12,8 +12,6 @@ File Class
 
 ----
 """
-from __future__ import absolute_import
-
 from ogs5py.fileclasses.pcs.core import PCS
 
 __all__ = ["PCS"]