+++ ogs5py/fileclasses/rfd/__init__.py
@@ -12,8 +12,6 @@ File Class
 
 ----
 """
-from __future__ import absolute_import
-
 from ogs5py.fileclasses.rfd.core import RFD
 
 __all__ = ["RFD"]