+++ ogs5py/tools/download.py
@@ -17,19 +17,11 @@ A downloading routine to get the OSG5 executable.
 
 ----
 """
-from __future__ import division, print_function, absolute_import
-
 import os
 import shutil
 import tarfile
 import zipfile
-
-try:  # PY3
-    from urllib.request import urlretrieve, urlopen
-except ImportError:  # PY2
-    from urllib import urlretrieve
-    from urllib2 import urlopen
-
+from urllib.request import urlretrieve, urlopen
 import tempfile
 import platform
 import lxml.html