+++ CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to **ogs5py** will be documented in this file.
 
 
+## [1.1.0] - 2020-03-22
+
+### Bugfixes
+* meshio 4 was not compatible
+* fixed integer type in exporting meshes with element/material IDs
+* better check for OGS5 success on Windows
+
+### Changes
+* drop py2.7 support
+
+
 ## [1.0.5] - 2019-11-18
 
 ### Bugfixes